mrc-leaf-helpers(extract-apply-torch-tp-move,mechanical_provable): Move apply_torch_tp to layers.model_parallel (cut+paste)
